Create DB struct and class for mock data using H2
It was created a file schema.sql to create the databe schema and a file
data.sql to simulate the insert of data into this h2 db.
It was created a DataSourceConfig class to setup and pont the domains
and repositories packages.
The application.yml file ws modified to setup the use of h2 database and
datasource.